requirements.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
idna==2.10
install==1.3.3
psycopg2-binary==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.18
twilio==6.44.1
urllib3==1.25.10
